split cli.ts composition root into src/pipeline/composition.ts
cli.ts was owning four unrelated responsibilities in a file whose
name implies only the last one: hardcoded body config, env binding,
layer wiring, and CLI command shapes. Move the first three (and the
placeholder PDF extractor) into composition.ts so cli.ts is strictly
about command definitions and NodeRuntime.runMain.

composition.ts exports:
- DEFAULT_BODIES (hardcoded governing-body list) + BodyConfigEntry type
- placeholderPdfExtract (follow-up slice will replace)
- readEnv / requireEnv helpers
- buildProductionLayers (the real composition root)

cli.ts loses: the Turso/Resend/Drizzle/service Live imports, the
body list, the env helpers, and buildProductionLayers. It keeps:
dotenv loading, the three Options declarations, runCommand,
listBodiesCommand, rootCommand, and the Command.run + runMain call.

Cold boot verified: `pnpm pipeline list-bodies` loads the .env file,
imports composition.ts, and prints the four bodies. All 91 tests
still pass; there was no test coverage of cli.ts or composition.ts
(both are production wiring code, not testable in isolation without
mocking every SDK).
