+++ src/pipeline/composition.ts
@@ -1,0 +1,177 @@
+import { createClient } from "@libsql/client";
+import { drizzle } from "drizzle-orm/libsql";
+import { Layer } from "effect";
+import { Resend } from "resend";
+import * as schema from "#/db/schema.ts";
+import { AlertServiceLive } from "#/pipeline/services/AlertService.ts";
+import { FinalsiteScraperLive } from "#/pipeline/services/FinalsiteScraper.ts";
+import { createGeminiSummarizer } from "#/pipeline/services/GeminiSummarizer.ts";
+import { EgovScraperLive } from "#/pipeline/services/ScraperService.ts";
+import { StorageServiceLive } from "#/pipeline/services/StorageService.ts";
+import { SummarizationServiceLive } from "#/pipeline/services/SummarizationService.ts";
+import { TranscriptionServiceLive } from "#/pipeline/services/TranscriptionService.ts";
+import { YouTubeScraperLive } from "#/pipeline/services/YouTubeScraper.ts";
+
+/**
+ * Effect teaching note: This file is the composition root — the single place
+ * where every real service layer is constructed with production config and
+ * merged into one big Layer that the orchestrator runs against. Everything
+ * below the CLI boundary (orchestrator, services) has never heard of Resend,
+ * Turso, or the YouTube API; they only see their injected dependencies.
+ *
+ * Separating the composition root from cli.ts means the CLI file owns only
+ * command shapes and `NodeRuntime.runMain` wiring — the layer graph, the
+ * body list, and the env binding all live here where they can be read and
+ * maintained without navigating past `@effect/cli` Option/Command boilerplate.
+ */
+
+/**
+ * A governing-body entry in the hardcoded list. Exactly one source field
+ * should be present per body — the union is kept open-ended here rather
+ * than via a discriminated union because bodies may eventually aggregate
+ * multiple sources (e.g. a body with both an eGov listing and a YouTube
+ * playlist).
+ */
+type BodyConfigEntry = {
+	slug: string;
+	name: string;
+	egovSearchType?: string;
+	finalsiteUrl?: string;
+	youtubePlaylistId?: string;
+};
+
+/**
+ * Hardcoded governing body list for this initial CLI. Later iterations should
+ * read this from the `governing_bodies` table so adding a body is a DB insert,
+ * not a code change — but for the first end-to-end run, hardcoding keeps the
+ * slice small.
+ */
+const DEFAULT_BODIES: BodyConfigEntry[] = [
+	{
+		slug: "ellettsville-town-council",
+		name: "Ellettsville Town Council",
+		egovSearchType: "12",
+	},
+	{
+		slug: "ellettsville-plan-commission",
+		name: "Ellettsville Plan Commission",
+		egovSearchType: "12",
+	},
+	{
+		slug: "monroe-county-commissioners",
+		name: "Monroe County Commissioners",
+		egovSearchType: "12",
+	},
+	{
+		slug: "richland-bean-blossom-school-board",
+		name: "Richland-Bean Blossom School Board",
+		finalsiteUrl: "https://www.rbbschools.net/school-board",
+	},
+];
+
+/**
+ * Placeholder PDF extractor. Production wiring should swap this for a real
+ * extractor (pdf-parse, unpdf, pdfjs-dist). Kept inline here so the CLI
+ * runs end-to-end in dry-run mode without adding a new dependency as part
+ * of this slice — the follow-up PDF-extraction slice will replace it.
+ */
+async function placeholderPdfExtract(bytes: ArrayBuffer): Promise<string> {
+	return `[PDF placeholder — ${bytes.byteLength} bytes; install a PDF extractor to see real text]`;
+}
+
+function readEnv(name: string): string | undefined {
+	const value = process.env[name];
+	return value && value.length > 0 ? value : undefined;
+}
+
+function requireEnv(name: string): string {
+	const value = readEnv(name);
+	if (!value) {
+		throw new Error(
+			`Missing required environment variable: ${name}. Set it in .env.local or the shell.`,
+		);
+	}
+	return value;
+}
+
+type BuildLayersInput = { dryRun: boolean };
+
+function buildProductionLayers(input: BuildLayersInput) {
+	const databaseUrl = requireEnv("DATABASE_URL");
+	const authToken = readEnv("TURSO_AUTH_TOKEN");
+
+	const client = createClient({
+		url: databaseUrl,
+		...(authToken ? { authToken } : {}),
+	});
+	const db = drizzle(client, { schema });
+
+	const geminiModelId = readEnv("GEMINI_MODEL") ?? "gemini-2.5-flash";
+	const geminiGenerator = createGeminiSummarizer({ modelId: geminiModelId });
+
+	const resendApiKey = input.dryRun ? undefined : readEnv("RESEND_API_KEY");
+	const resendFrom =
+		readEnv("RESEND_FROM") ?? "civic-mirror@alerts.example.com";
+	const resendTo = readEnv("ALERT_EMAIL") ?? "ops@example.com";
+
+	const resendClient = resendApiKey ? new Resend(resendApiKey) : null;
+
+	const egovBaseUrl =
+		readEnv("EGOV_BASE_URL") ??
+		"https://ellettsville.in.us/egov/apps/document/center.egov";
+
+	const egov = EgovScraperLive({ baseUrl: egovBaseUrl });
+
+	const finalsite = FinalsiteScraperLive({
+		baseUrl: readEnv("FINALSITE_BASE_URL") ?? "https://www.rbbschools.net",
+	});
+
+	const youtube = YouTubeScraperLive({
+		apiKey: readEnv("YOUTUBE_API_KEY") ?? "missing-key",
+	});
+
+	const transcription = TranscriptionServiceLive();
+
+	const summarization = SummarizationServiceLive({
+		model: geminiModelId,
+		generateFn: geminiGenerator,
+	});
+
+	const storage = StorageServiceLive(db);
+
+	const alert = AlertServiceLive({
+		to: resendTo,
+		from: resendFrom,
+		sendFn: resendClient
+			? async ({ to, from, subject, body }) => {
+					await resendClient.emails.send({
+						to,
+						from,
+						subject,
+						text: body,
+					});
+				}
+			: async ({ subject, body }) => {
+					console.log(`[alert-stub] ${subject}\n${body}`);
+				},
+	});
+
+	return Layer.mergeAll(
+		egov,
+		finalsite,
+		youtube,
+		transcription,
+		summarization,
+		storage,
+		alert,
+	);
+}
+
+export {
+	DEFAULT_BODIES,
+	placeholderPdfExtract,
+	readEnv,
+	requireEnv,
+	buildProductionLayers,
+};
+export type { BodyConfigEntry, BuildLayersInput };
